Atualizando ContaService

+++ src/main/java/com/core_banking/DTO/ContaDTO.java
@@ -1,10 +1,14 @@
 package com.core_banking.DTO;
 
+import java.math.BigDecimal;
+
 public class ContaDTO {
 
     private Integer numeroConta;
     private String titularConta;
     private String tipoConta;
-    private Double saldo;
-    private Double limite;
+    private BigDecimal saldo;
+    private BigDecimal limite;
+
+
 }
